+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added support for dates in `DD.MM.YYYY` format in the activities import
 - Set up the language localization for Türkçe (`tr`)
 
 ## 2.4.0 - 2023-09-19